Key restructuring changes, now use three layers of abstraction.
