--- local_data/positive_examples.tsv
@@ -1,7 +1,0 @@
-curator	date	ev_json	id	pa_hash	pa_json	source	source_hash	tag	text	english	agent_json_list	synonyms_list	agent_info
-bachmanjohn@gmail.com	Thu, 21 May 2020 00:00:11 GMT		3564	33884862164561731		bioexp_paper_rlimsp	8770739897070071093	correct	Bottom, quantified PAK4 autophosphorylation relative to the zero time point of PAK4 autophosphorylation are indicated below.	PAK4 phosphorylates itself.	[{'name': 'PAK4', 'db_refs': {'UP': 'O96013', 'HGNC': '16059', 'TEXT': 'PAK4'}}]		{'HGNC:16059': {'name': 'PAK4', 'synonyms': ['KIAA1142', 'PAK-4', 'PAK4', 'Serine/threonine-protein kinase PAK 4', 'p21 (RAC1) activated kinase 4', 'p21-activated kinase 4'], 'definition': None, 'syn_in_text': 'PAK4', 'syn_in_stmt': 'PAK4'}}
-ben.gyori@gmail.com	Wed, 20 May 2020 14:17:21 GMT		3142	-20338582019148488		bioexp_paper_rlimsp	2334023753075164564	correct	Protection of CDC6 by CUL4B is dependent on CDC6 phosphorylation by CDK2.	CDK2 phosphorylates CDC6.	[{'name': 'CDK2', 'db_refs': {'UP': 'P24941', 'HGNC': '1771', 'TEXT': 'Cdk2'}}, {'name': 'CDC6', 'db_refs': {'UP': 'Q99741', 'HGNC': '1744', 'TEXT': 'Cdc6'}}]		{'HGNC:1771': {'name': 'CDK2', 'synonyms': ['CDK2', 'CDK2 Protein', 'CDKN2', 'Cdc2 Related Protein Kinase', 'Cdc2-Related Protein Kinase', 'Cdk2', 'Cdk2 Protein Kinase', 'Cell division protein kinase 2', 'Cyclin Dependent Kinase 2', 'Cyclin-Dependent Kinase 2', 'Cyclin-dependent kinase 2', 'cdk2', 'cyclin dependent kinase 2', 'p33 protein kinase', 'p33cdk2 Kinase'], 'definition': None, 'syn_in_text': 'CDK2', 'syn_in_stmt': 'CDK2'}, 'HGNC:1744': {'name': 'CDC6', 'synonyms': ['CDC18L', 'CDC6', 'CDC6-related protein', 'Cdc18-related protein', 'Cdc6', 'Cell division control protein 6 homolog', 'HsCDC6', 'HsCdc18', 'cell division cycle 6', 'p62(cdc6)'], 'definition': None, 'syn_in_text': 'CDC6', 'syn_in_stmt': 'CDC6'}}
-bachmanjohn@gmail.com	Wed, 20 May 2020 14:43:19 GMT		3171	867775915671121		bioexp_paper_rlimsp	401068084514771088	correct	Using this antibody, we showed that YAP S61 phosphorylation increased following 2-DG treatment in wild-type mouse embryonic fibroblasts (MEFs) but not in AMPK-knockout MEFs (Figure 4E), suggesting that AMPK phosphorylates YAP at the S61 site in response to energy stress.	AMPK leads to the phosphorylation of YAP1 on S61.	[{'name': 'AMPK', 'db_refs': {'FPLX': 'AMPK', 'TEXT': 'AMP-activated protein kinase'}}, {'name': 'YAP1', 'db_refs': {'UP': 'P46937', 'HGNC': '16262', 'TEXT': 'YAP'}}]		{'FPLX:AMPK': {'name': 'AMPK', 'synonyms': ['AMP-activated', 'AMPK', 'AMP-activated protein kinase'], 'definition': None, 'syn_in_text': 'AMPK', 'syn_in_stmt': 'AMPK'}, 'HGNC:16262': {'name': 'YAP1', 'synonyms': ['Protein yorkie homolog', 'Transcriptional coactivator YAP1', 'YAP', 'YAP-1', 'YAP1', 'YAP65', 'Yes-associated protein 1', 'Yes-associated protein YAP65 homolog', 'Yes1 associated transcriptional regulator'], 'definition': None, 'syn_in_text': 'YAP', 'syn_in_stmt': 'YAP1'}}
-ben.gyori@gmail.com	Fri, 19 Feb 2021 01:07:02 GMT		11067	-32671237630440129		bioexp_paper_sparser	7459679528160511391	correct	In the present study we employed tandem affinity purification method to screen for IP6K interaction partners and found that human IP6K2 directly binds to Tti1.	IP6K2 binds TTI1.	[{'name': 'IP6K2', 'db_refs': {'UP': 'Q9UHH9', 'HGNC': '17313', 'TEXT': 'IP6K2'}}, {'name': 'TTI1', 'db_refs': {'UP': 'O43156', 'HGNC': '29029', 'TEXT': 'Tti1'}}]		{'HGNC:17313': {'name': 'IP6K2', 'synonyms': ['IHPK2', 'IP6K2', 'Inositol hexakisphosphate kinase 2', 'InsP6 kinase 2', 'InsP6K2', 'P(i)-uptake stimulator', 'PiUS', 'inositol hexakisphosphate kinase 2'], 'definition': None, 'syn_in_text': 'IP6K2', 'syn_in_stmt': 'IP6K2'}, 'HGNC:29029': {'name': 'TTI1', 'synonyms': ['KIAA0406', 'Protein SMG10', 'SMG10', 'TELO2 interacting protein 1', 'TELO2-interacting protein 1 homolog', 'TTI1', 'smg-10', 'smg-10 homolog, nonsense mediated mRNA decay factor (C. elegans)', 'Tti1'], 'definition': None, 'syn_in_text': 'TTI1', 'syn_in_stmt': 'TTI1'}}
-ben.gyori@gmail.com	Fri, 09 Apr 2021 18:03:54 GMT		11761	26908342305891100		bioexp_paper_rlimsp	-1917340880992183075	correct	Lck undergoes autophosphorylation on both Y394 and Y505 (a) Immunoblot analysis of Lck autophosphorylation on membranes versus in solution using WB.	LCK phosphorylates itself on Y505.	[{'name': 'LCK', 'db_refs': {'UP': 'P06239', 'HGNC': '6524', 'TEXT': 'Lck'}}]		{'HGNC:6524': {'name': 'LCK', 'synonyms': ['LCK', 'LCK proto-oncogene, Src family tyrosine kinase', 'LSK', 'Lck', 'Leukocyte C-terminal Src kinase', 'Lymphocyte cell-specific protein-tyrosine kinase', 'Protein YT16', 'Proto-oncogene Lck', 'T cell-specific protein-tyrosine kinase', 'Tyrosine-protein kinase Lck', 'p56', 'p56-LCK'], 'definition': None, 'syn_in_text': 'LCK', 'syn_in_stmt': 'LCK'}}
-ben.gyori@gmail.com	Mon, 06 Jan 2020 05:56:44 GMT		1020	-55474658534404		bioexp_paper_reach	8741211897633255399	correct	PKCdelta dependent phosphorylation of p21 Cip1 and WAF1 at Ser146 regulates its subcellular distribution and function.	PRKCD phosphorylates CDKN1A on S146.	[{'name': 'PRKCD', 'db_refs': {'UP': 'Q05655', 'HGNC': '9399', 'TEXT': 'PKCδ'}}, {'name': 'CDKN1A', 'db_refs': {'UP': 'P38936', 'HGNC': '1784', 'TEXT': 'p21'}}]		{'HGNC:9399': {'name': 'PRKCD', 'synonyms': ['Kinase C-delta, Protein', 'PKC delta Serine Threonine Kinase', 'PKC-delta Serine-Threonine Kinase', 'PKCD', 'PKCdelta', 'PKCδ', 'PRKCD', 'Protein Kinase C delta', 'Protein Kinase C-delta', 'Protein kinase C delta type', 'Serine-Threonine Kinase, PKC-delta', 'Tyrosine-protein kinase PRKCD', 'nPKC-delta', 'protein kinase C delta'], 'definition': None, 'syn_in_text': 'PKCdelta', 'syn_in_stmt': 'PRKCD'}, 'HGNC:1784': {'name': 'CDKN1A', 'synonyms': ['CAP20', 'CDK-interacting protein 1', 'CDK2 Associated Protein 20 kDa', 'CDK2-Associated Protein 20 kDa', 'CDKN1', 'CDKN1 Protein', 'CDKN1A', 'CDKN1A Protein', 'CIP1', 'Cdk Interacting Protein 1', 'Cdk-Interacting Protein 1', 'Cdk2 Inhibitor Protein', 'Cell Cycle Regulator p21', 'Cyclin Dependent Kinase Inhibitor 1A Protein', 'Cyclin Dependent Kinase Inhibitor p21', 'Cyclin Kinase Inhibitor p21', 'Cyclin-Dependent Kinase Inhibitor 1A Protein', 'Cyclin-Dependent Kinase Inhibitor p21', 'Cyclin-dependent kinase inhibitor 1', 'MDA-6', 'MDA6', 'Melanoma differentiation-associated protein 6', 'P21', 'PIC1', 'SDI1', 'Senescent Cell Derived Inhibitor Protein 1', 'Senescent Cell-Derived Inhibitor Protein 1', 'WAF1', 'cyclin dependent kinase inhibitor 1A', 'p21', 'p21 Cell Cycle Regulator', 'p21 Cip1', 'p21 Cyclin Kinase Inhibitor', 'p21 WAF1', 'p21CIP1', 'p21Cip1/Waf1'], 'definition': None, 'syn_in_text': 'CIP1', 'syn_in_stmt': 'CDKN1A'}}
