--- local_data/sample_curation.json
@@ -1,1 +1,0 @@
-[{"curator": "ben.gyori@gmail.com", "date": "Wed, 07 Oct 2020 18:56:32 GMT", "ev_json": null, "id": 6547, "pa_hash": -34382679073582559, "pa_json": null, "source": "bioexp_paper_medscan", "source_hash": 693729828248637321, "tag": "correct"}, {"curator": "ben.gyori@gmail.com", "date": "Wed, 07 Oct 2020 18:56:14 GMT", "ev_json": null, "id": 6545, "pa_hash": -34382679073582559, "pa_json": null, "source": "bioexp_paper_medscan", "source_hash": 8577423735471948069, "tag": "correct"}, {"curator": "bachmanjohn@gmail.com", "date": "Tue, 12 May 2020 14:50:25 GMT", "ev_json": null, "id": 2787, "pa_hash": 28854425604630862, "pa_json": null, "source": "bioexp_paper_rlimsp", "source_hash": 2750584666843918, "tag": "correct"}, {"curator": "bachmanjohn@gmail.com", "date": "Wed, 10 Jun 2020 14:47:45 GMT", "ev_json": null, "id": 4730, "pa_hash": 30681958885786940, "pa_json": null, "source": "bioexp_paper_multi", "source_hash": -1534389185697378360, "tag": "correct"}, {"curator": "ben.gyori@gmail.com", "date": "Wed, 07 Oct 2020 19:14:01 GMT", "ev_json": null, "id": 6574, "pa_hash": -24711765246990842, "pa_json": null, "source": "bioexp_paper_medscan", "source_hash": 6361324680027062954, "tag": "correct"}, {"curator": "bachmanjohn@gmail.com", "date": "Wed, 10 Jun 2020 14:48:44 GMT", "ev_json": null, "id": 4732, "pa_hash": 30681958885786940, "pa_json": null, "source": "bioexp_paper_multi", "source_hash": 2413375295660655511, "tag": "correct"}, {"curator": "ben.gyori@gmail.com", "date": "Wed, 07 Oct 2020 18:56:22 GMT", "ev_json": null, "id": 6546, "pa_hash": -34382679073582559, "pa_json": null, "source": "bioexp_paper_medscan", "source_hash": 7064910908075962224, "tag": "correct"}, {"curator": "bachmanjohn@gmail.com", "date": "Tue, 12 May 2020 14:47:03 GMT", "ev_json": null, "id": 2782, "pa_hash": 30681958885786940, "pa_json": null, "source": "bioexp_paper_rlimsp", "source_hash": 1142110265793306722, "tag": "mod_site"}, {"curator": "ben.gyori@gmail.com", "date": "Wed, 07 Oct 2020 19:13:40 GMT", "ev_json": null, "id": 6571, "pa_hash": -24711765246990842, "pa_json": null, "source": "bioexp_paper_medscan", "source_hash": 2411624995189796859, "tag": "correct"}, {"curator": "ben.gyori@gmail.com", "date": "Wed, 07 Oct 2020 18:37:30 GMT", "ev_json": null, "id": 6532, "pa_hash": -35508525441903305, "pa_json": null, "source": "bioexp_paper_medscan", "source_hash": -9146469845369368878, "tag": "correct"}]
